requirements.txt
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.45.0
peft>=0.12.0
bitsandbytes>=0.43.0
accelerate>=0.26.0
qwen-vl-utils>=0.0.8
Pillow>=10.0.0
numpy>=1.24.0
tqdm>=4.65.0
pyyaml>=6.0
python-Levenshtein>=0.23.0
